backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary       # PostgreSQL driver
python-dotenv
pydantic
httpx